Split the secret's lib.rs file
