+++ engine/vault/.license_template
@@ -1,0 +1,1 @@
+../.license_template
